requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
